SMGR: reduce number of output files and update doxygen
* studymanager.log now contains logs of SMGR
* run_case.log in DEST/STUDY/CASE/RESU/run_id contains logs of code_saturne run command
* report_global is not generated anymore
* report_figures is the list of generated figures

+++ python/code_saturne/studymanager/cs_studymanager_texmaker.py
@@ -39,11 +39,10 @@ from code_saturne.studymanager.cs_studymanager_run import run_studymanager_comma
 class TexWriter(object):
     """
     """
-    def __init__(self, dest, filename, log, pdflatex):
+    def __init__(self, dest, filename, pdflatex):
         self.__dest = dest
         self.__filename = os.path.join(self.__dest, filename)
         self.__doc = []
-        self.__log = log
         self.__pdflatex = pdflatex
 
 
@@ -168,15 +167,23 @@ class TexWriter(object):
 
     def make_pdf(self):
         """
-        Buld the pdf file, and clean the temporary files.
+        Build the pdf file, and clean the temporary files.
         """
         cmd = "pdflatex " + self.__filename + ".tex"
-        r, t = run_studymanager_command(cmd, self.__log)
+        log_name = "make_pdf.log"
+        log_file = open(log_name, mode='w')
+        error, time = run_studymanager_command(cmd, log_file)
+        log_file.close()
 
-        for suffixe in ["tex", "log", "aux"]:
-            f = self.__filename + "." + suffixe
-            if os.path.isfile(f):
-                os.remove(self.__filename + "." + suffixe)
+        if not error:
+            os.remove(log_name)
+            for suffixe in ["tex", "log", "aux"]:
+                f = self.__filename + "." + suffixe
+                if os.path.isfile(f):
+                    os.remove(self.__filename + "." + suffixe)
+        else:
+            print(" /!\ ERROR during pdf generation. See %s\n",
+                  log_name)
 
         return self.__filename + ".pdf"
 
@@ -195,59 +202,14 @@ class TexWriter(object):
 
 #-------------------------------------------------------------------------------
 
-class Report1(TexWriter):
+class Report(TexWriter):
     """
-    Global report.
+    Figures report.
     """
-    def __init__(self, dest, label, log, report, xml, pdflatex):
-        TexWriter.__init__(self, dest, label, log, pdflatex)
-        self.appendLine("\\section{Summary}")
-        self.tabCreate(["Study / Case", "Compilation", "Run", "Time (s)", "Difference"])
-        self.xml = xml
-        self.report = report
-
-
-    def add_row(self, studyLabel, caseLabel, is_compil, is_run, is_time, is_compare, is_diff):
-        if is_compare == "not done":
-            threshold = "Not used"
-            is_diff   = "Not used"
-
-        label = "%s / %s" % (studyLabel, caseLabel)
-        label = label.replace("_", "\_")
-        self.tabWrite([label, is_compil, is_run, is_time, is_diff])
-
-
-    def close(self):
-        self.tabClose()
-
-        self.appendLine("\\section{Log}")
-        self.appendLine("\\tiny\n\\begin{verbatim}")
-        f = open(self.report)
-        self.rawLine(f.read())
-        f.close()
-        self.appendLine("\\end{verbatim}\n\\normalsize")
-
-        self.appendLine("\\section{File of commands}")
-        self.appendLine("\\tiny\n\\begin{verbatim}")
-        self.rawLine(self.xml)
-        self.appendLine("")
-        self.appendLine("\\end{verbatim}\n\\normalsize")
-
-        self.write()
-
-        return self.tex_finalize()
-
-#-------------------------------------------------------------------------------
-
-class Report2(TexWriter):
-    """
-    Detailed report.
-    """
-    def __init__(self, dest, label, log, pdflatex):
+    def __init__(self, dest, label, pdflatex):
         """
         """
-        TexWriter.__init__(self, dest, label, log, pdflatex)
-
+        TexWriter.__init__(self, dest, label, pdflatex)
 
     def add_row(self, values, studyLabel, caseLabel):
         nbvalue = len(values)
@@ -265,95 +227,5 @@ class Report2(TexWriter):
     def close(self):
         self.write()
         return self.tex_finalize()
-
-#-------------------------------------------------------------------------------
-
-def test():
-    """
-    Test function.
-    The pupose is to build 2 documents:
-    1. a global report for all cases:
-    2. a detailled report for each case:
-    """
-
-    dest      = os.getcwd()
-    doc       = Report1(dest, "r1")
-    compil    = True
-    run       = True
-    compare   = True
-    diff      = True
-    repo      = True
-    threshold = 1.e-15
-    doc.add_row("MYSTUDY", "MYCASE", compil, run, compare, diff, repo, threshold)
-    doc.close()
-
-    msg = """
-  .----------------------------.
-  |   code_saturne file dump   |
-  `----------------------------'
-
-Opening input file: "RESU/20110217-2231/checkpoint/main"
-
-  File type: Checkpoint / restart, R0
-
-  Base header size: 128
-  Header alignment: 64
-  Body alignment:   64
-
-Opening input file: "RESU/20110217-2233/checkpoint/main"
-
-  File type: Checkpoint / restart, R0
-
-  Base header size: 128
-  Header alignment: 64
-  Body alignment:   64
-
-  "nbre_pas_de_temps               "; Location:  0; Type: i4    ; Size: 1
-    Differences: 1
-
-  "instant_precedent               "; Location:  0; Type: r8    ; Size: 1
-    Differences: 1; Max: 200; Mean: 200
-
-  "pression_ce_phase01             "; Location:  1; Type: r8    ; Size: 14574
-    Differences: 14574; Max: 3.09509; Mean: 0.159934
-
-  "vitesse_u_ce_phase01            "; Location:  1; Type: r8    ; Size: 14574
-    Differences: 14574; Max: 2.00929; Mean: 0.118064
-
-  "vitesse_v_ce_phase01            "; Location:  1; Type: r8    ; Size: 14574
-    Differences: 14574; Max: 1.84415; Mean: 0.065676
-
-  "vitesse_w_ce_phase01            "; Location:  1; Type: r8    ; Size: 14574
-    Differences: 14574; Max: 1.15005; Mean: 0.093984
-
-  "k_ce_phase01                    "; Location:  1; Type: r8    ; Size: 14574
-    Differences: 14574; Max: 2.49588; Mean: 0.0191537
-
-  "eps_ce_phase01                  "; Location:  1; Type: r8    ; Size: 14574
-    Differences: 14574; Max: 7438.13; Mean: 12.6477
-
-  "scalaire_ce_0001                "; Location:  1; Type: r8    ; Size: 14574
-    Differences: 14500; Max: 0.11862; Mean: 0.00345369
-
-    """
-
-    doc  = Report2(dest, "r2")
-    v    = []
-    info = msg.replace("\"", " ").replace(";", " ").replace(":", " ").split()
-    print(info)
-    repo = "RESU/20110217-2231/checkpoint/main"
-    dest = "RESU/20110217-2233/checkpoint/main"
-
-    for i in range(len(info)):
-        if info[i][:4] == 'Diff':
-            if info[i-3] not in ['i4', 'u4']:
-                v.append([info[i-7].replace("_", "\_"), info[i+3], info[i+5]])
-    print(v)
-    doc.add_row(v, "MYSTUDY", "MYCASE", threshold)
-    doc.close()
-
-
-if __name__ == '__main__':
-    test()
 
 #-------------------------------------------------------------------------------
